refactor(profit): 변경된 로직에 맞춰 구조 변경

+++ src/main/java/blackjack/view/dto/TotalWinningResult.java
@@ -3,7 +3,7 @@ package blackjack.view.dto;
 import java.util.List;
 
 public record TotalWinningResult(
-        int dealerProfit,
+        long dealerProfit,
         List<PlayerProfit> playerResults
 ) {
 }
